+++ src/main.rs
@@ -1,4 +1,5 @@
 use clap::Parser;
+use tokio::runtime::Builder;
 
 mod cli;
 mod commands;
@@ -9,7 +10,12 @@ fn main() -> anyhow::Result<()> {
 
     match opts.command {
         cli::Command::Inspect(command) => command.run(),
-        cli::Command::Backup(cli::Backup::Verify(command)) => command.run(),
+        cli::Command::Backup(cli::Backup::Verify(command)) => {
+            let runtime = Builder::new_multi_thread()
+                .thread_name("tev-worker")
+                .build()?;
+            runtime.block_on(command.run())
+        }
         cli::Command::Backup(cli::Backup::Mount(command)) => command.run(),
     }
 }